add container, slot, and grid class
